[PEAR-5] rename to user/entity
