+++ src/main/java/com/dsmpear/main/domain/auth/service/AuthServiceImpl.java
@@ -5,8 +5,8 @@ import com.dsmpear.main.domain.auth.entity.refreshtoken.RefreshTokenRepository;
 import com.dsmpear.main.domain.auth.dto.request.SignInRequest;
 import com.dsmpear.main.domain.auth.dto.response.TokenResponse;
 import com.dsmpear.main.domain.auth.exceptions.UserNotFoundException;
-import com.dsmpear.main.domain.user.domain.User;
-import com.dsmpear.main.domain.user.domain.UserRepository;
+import com.dsmpear.main.domain.user.entity.User;
+import com.dsmpear.main.domain.user.entity.UserRepository;
 import com.dsmpear.main.global.security.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
