+++ src/main/java/com/dsmpear/main/global/security/auth/AuthDetails.java
@@ -1,6 +1,6 @@
 package com.dsmpear.main.global.security.auth;
 
-import com.dsmpear.main.domain.user.domain.User;
+import com.dsmpear.main.domain.user.entity.User;
 import lombok.AllArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
